Improve place/remove tile packet accuracy

+++ NetworkGame/Packets/PlaceTilePacket.cs
@@ -1,15 +1,15 @@
 public class PlaceTilePacket : Packet
 {
-    public int X;
-    public int Y;
+    public float X;
+    public float Y;
     public int TextureId;
 
     public PlaceTilePacket(byte[]? data)
     {
         if (data == null) return;
         if (data[0] != PacketId.PACKET_PLACE_TILE) throw new InvalidDataException($"Invalid Packet Id; Got {data[0]} expected {PacketId.PACKET_PLACE_TILE}");
-        X = BitConverter.ToInt32(data[1..5]);
-        Y = BitConverter.ToInt32(data[5..9]);
+        X = BitConverter.ToSingle(data[1..5]);
+        Y = BitConverter.ToSingle(data[5..9]);
         TextureId = BitConverter.ToInt32(data[9..13]);
     }
