+++ NetworkGame/Packets/RemoveTilePacket.cs
@@ -1,14 +1,14 @@
 public class RemoveTilePacket : Packet
 {
-    public int X;
-    public int Y;
+    public float X;
+    public float Y;
 
     public RemoveTilePacket(byte[]? data)
     {
         if (data == null) return;
         if (data[0] != PacketId.PACKET_REMOVE_TILE) throw new InvalidDataException($"Invalid Packet Id; Got {data[0]} expected {PacketId.PACKET_REMOVE_TILE}");
-        X = BitConverter.ToInt32(data[1..5]);
-        Y = BitConverter.ToInt32(data[5..9]);
+        X = BitConverter.ToSingle(data[1..5]);
+        Y = BitConverter.ToSingle(data[5..9]);
     }
     
     public override byte[] Serialize()
